study_bot_project/requirements.txt
fastapi
uvicorn
langchain
langchain_groq
pymongo
python-dotenv